fixed the bugs in editing function
Also made it so that the trait box that is currently being edited is
hidden to avoid confusion. Last commit had a typo- in order to see the
full flow, you must select a political party- since I still need to do
validation for that and redirect someone to the free-response identity
section if they choose not to select a party.